Release 4.0.0

+++ Rakefile
@@ -176,7 +176,9 @@ task :create_desktop_cre do
   publicancre = "#{home}/.local/share/applications/publicancreators.desktop"
   publicancreico = File.expand_path(File.join(File.dirname(__FILE__), 'bin/', 'publican.png'))
   publicancrebin = File.expand_path(File.join(File.dirname(__FILE__), 'bin/', 'PublicanCreators.rb'))
-  FileUtils.rm(publicancre)
+  if File.exists?(publicancre)
+    FileUtils.rm(publicancre)
+  end
   puts 'Creating Desktop file for PublicanCreators'.color(:yellow)
   FileUtils.touch "#{publicancre}"
   File.write "#{publicancre}", <<EOF
@@ -197,7 +199,9 @@ task :create_desktop_rev do
   publicanrev = "#{home}/.local/share/applications/publicancreators-rev.desktop"
   publicanrevico = File.expand_path(File.join(File.dirname(__FILE__), 'bin/', 'publican-revision.png'))
   publicanrevbin = File.expand_path(File.join(File.dirname(__FILE__), 'bin/', 'RevisionCreator.rb'))
-  FileUtils.rm(publicanrev)
+  if File.exist?(publicanrev)
+    FileUtils.rm(publicanrev)
+  end
   puts 'Creating Desktop file for PublicanCreatorsRevision'.color(:yellow)
   FileUtils.touch "#{publicanrev}"
   File.write "#{publicanrev}", <<EOF
